refactor table fix tests
